adjustment for code review

+++ ToysAndGamesBusiness/Services/ProductService.cs
@@ -1,39 +1,36 @@
 ﻿using StockManagementEntities.Models;
 using StockManagementPersistence;
-using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
-using System.Text;
-using System.Threading.Tasks;
+using ToysAndGamesServices.Contracts;
 using ToysAndGamesUtil;
 
-namespace ToysAndGamesBusiness
+namespace ToysAndGamesBusiness.Services
 {
-    public class ProductBusiness : IProductBusiness
+    public class ProductService : IProductService
     {
         private readonly ToysAndGamesDbContext _db;
         private readonly ILocalStorage _localSt;
 
-        public ProductBusiness(ToysAndGamesDbContext db, ILocalStorage localSt)
+        public ProductService(ToysAndGamesDbContext db, ILocalStorage localSt)
         {
             _db = db;
             _localSt = localSt;
         }
-        public Product CreateOrUpdate(Product product)
+        public Product UpSert(Product product)
         {
             try
             {
                 //TODO: ERRROR if productId its 0 its my product not going to be added with the 0 ID ?
-                if (!product.Id.HasValue)
+                //Comments: ID product is an identity column
+                if (product.Id <= 0)
                     _db.Products.Add(product);
                 else
                     _db.Products.Update(product);
 
                 _db.SaveChanges();
 
-                if (product.File != null)
-                    _localSt.StoreFile(product.File, product.Id.ToString());
+                if (product.File != null) _localSt.StoreFile(product.File, product.Id.ToString());
 
                 return product;
             }
@@ -59,16 +56,14 @@ namespace ToysAndGamesBusiness
         {
             Product? product = null;
 
-            if (id == 0)
-                throw new Exception("Id Can't be null or 0");
-
             try
             {
-                product = _db.Products.FirstOrDefault(p => p.Id == id);
-                if (product == null)
-                {
-                    throw new Exception("Product Not Found");
-                }
+                if (id <= 0) throw new Exception("Id Can't be null or 0");
+
+                product = Get(p => p.Id == id).FirstOrDefault();
+
+                if (product == null) throw new Exception("Product Not Found");
+                
                 return product;
             }
             catch (Exception)
@@ -76,24 +71,17 @@ namespace ToysAndGamesBusiness
                 throw;
             }
         }
-
-        //public async Task<IList<Product>> Get(Expression<Func<Product, bool>> predicate, Expression<Func<Product)
-        //{
-        //    return _db.Products.Where(predicate).ToList();
-        //}
-
         public void Delete(int id)
         {
             Product? product = null;
 
-            if (id == 0)
-                throw new Exception("Id Can't be null or 0");
-
             try
             {
-                product = _db.Products.FirstOrDefault(p => p.Id == id);
-                if (product == null)
-                    throw new Exception("Product Not Found");
+                if (id <= 0)throw new Exception("Id Can't be null or 0");
+
+                product = Get(p => p.Id == id).FirstOrDefault();
+
+                if (product == null) throw new Exception("Product Not Found");
 
                 _db.Products.Remove(product);
                 _db.SaveChanges();
@@ -104,5 +92,9 @@ namespace ToysAndGamesBusiness
             }
         }
 
+        public List<Product> Get(Expression<Func<Product,bool>> predicate)
+        {
+            return _db.Products.Where(predicate).ToList();
+        }
     }
 }
